+++ lib/iron_admin.rb
@@ -3,6 +3,9 @@
 require "iron_admin/version"
 require "iron_admin/errors"
 require "iron_admin/configuration"
+require "iron_admin/live/poll_cache"
+require "iron_admin/live/broadcaster"
+require "iron_admin/live"
 require "iron_admin/adapters/base"
 require "iron_admin/adapters/registry"
 require "iron_admin/adapters/active_record"
@@ -17,6 +20,7 @@ require "iron_admin/nested_attributes_validator"
 require "iron_admin/concerns/nestable"
 require "iron_admin/concerns/soft_deletable"
 require "iron_admin/concerns/importable"
+require "iron_admin/concerns/live_updatable"
 require "iron_admin/filters/base_query_builder"
 require "iron_admin/filters/active_record_query_builder"
 require "iron_admin/filters/query_builder"
@@ -145,6 +149,7 @@ module IronAdmin
     def reset_configuration!
       @configuration = Configuration.new
       @dashboard_class = nil
+      Live.reset! if defined?(Live)
     end
 
     def register_field_type(type_name, &)